requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-celery-results==2.5.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
importlib-metadata==4.8.3
kombu==5.3.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
